Update Left Leaning Red Black Tree.py
changed return on insert to send node to fixup method to catch edge case where 2 leaves are red. This would violate rule 6 as shown on geeks for geeks: "There should not be a node which has LEFT RED child and RIGHT RED child, if present Invert the colors of all nodes i. e., current_node, LEFT child, and RIGHT child."

https://www.geeksforgeeks.org/left-leaning-red-black-tree-insertion/